travis: revised travis configuration
Multi OS support is a beta product and as a result it has to be requested and enabled for specific repositories. This makes it difficult for a user when they fork with the intention of submitting a PR. This change removes this dependency & instead just sets language to objective-c as this will ensure that an osx environment is used and since fsevents needs an osx environment and no other this results in the same behaviour.